getting most of the login and session stuff working

+++ app/models/User.scala
@@ -1,3 +1,3 @@
 package models
 
-case class User(username: String, password: String, id: Integer = 0)
+case class User(username: String, password: String, id: Long = 0)
